+++ CHANGELOG.md
@@ -22,7 +22,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Fixes
 
-- Fixes `Accordion` styles, including padding and active hover state on the accordion button.
+- Fixes `Accordion` styles, including padding and active hover state, and double-border issue.
 
 ## 3.5.4 (February 13, 2025)